Moves constants and removes unused type
