+++ src/phonemes/sign/sign-writing.ts
@@ -6,10 +6,6 @@
  * https://www.signbank.org/iswa/
  */
 
-export type SignWritingSymbol = {
-    symbol: string
-};
-
 export enum SignWritingCategory {
     HANDS = "Hands",
     MOVEMENT = "Movement",
